# Conflicts:
#	build.gradle
#	src/main/java/com/sparta/projectblue/config/SwaggerConfig.java
#	src/main/java/com/sparta/projectblue/domain/common/enums/ReservationStatus.java
#	src/main/java/com/sparta/projectblue/domain/common/service/TestService.java
#	src/main/java/com/sparta/projectblue/domain/reservation/entity/Reservation.java
#	src/main/java/com/sparta/projectblue/domain/reservation/repository/ReservationRepository.java
#	src/main/java/com/sparta/projectblue/domain/reservation/service/ReservationService.java

+++ src/main/java/com/sparta/projectblue/domain/hall/repository/HallRepository.java
@@ -3,5 +3,5 @@ package com.sparta.projectblue.domain.hall.repository;
 import com.sparta.projectblue.domain.hall.entity.Hall;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HallRepository extends JpaRepository<Hall, Integer> {
+public interface HallRepository extends JpaRepository<Hall, Long> {
 }
